move the command decorator to another file
